server/requirements.txt
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
pymysql>=1.1.0
python-dotenv>=1.0.1
pydantic>=2.7.1
